+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Gunpowder Market 1.0.2
+
+- Basically fix everything
+
+---
+
 # Gunpowder Market 1.0.1
 
 - Update to Gunpowder Core 0.3.2 (1.16.4)